feat(dal): add ability to individually specify builtin schemas
- Rename skip migrating schemas environment variable to indicate that
  it is for a list (or single item) of schema(s)
- Add documentation example for new builtin schema environment variable
  behavior
- Add "BuiltinSchemaOption" for all scenarios:
  - All: default behavior and able to specified
  - None: previous behavior of the environment variable (still possible)
  - Some: new behavior added to the environment variable
- Split builtin schemas into individual files on a "per schema" basis
  - This makes things more readable and helps the new environment
    variable behavior be easy to work with (rather than dealing with
    arbitrary nesting)
