+++ lib/dal-test/src/lib.rs
@@ -1,5 +1,6 @@
 //! This module provides helpers and resources for constructing and running integration tests.
 
+use std::collections::HashSet;
 use std::{
     borrow::Cow,
     env,
@@ -7,6 +8,7 @@ use std::{
     sync::{Arc, Once},
 };
 
+use dal::builtins::BuiltinSchemaOption;
 #[cfg(debug_assertions)]
 use dal::check_runtime_dependencies;
 use dal::{
@@ -46,7 +48,7 @@ const DEFAULT_PG_DBNAME: &str = "si_test";
 const ENV_VAR_NATS_URL: &str = "SI_TEST_NATS_URL";
 const ENV_VAR_PG_HOSTNAME: &str = "SI_TEST_PG_HOSTNAME";
 const ENV_VAR_PG_DBNAME: &str = "SI_TEST_PG_DBNAME";
-const ENV_VAR_SKIP_MIGRATING_SCHEMAS: &str = "SI_TEST_SKIP_MIGRATING_SCHEMAS";
+const ENV_VAR_BUILTIN_SCHEMAS: &str = "SI_TEST_BUILTIN_SCHEMAS";
 
 const JWT_PUBLIC_FILENAME: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/", "config/public.pem");
 const JWT_PRIVATE_FILENAME: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/", "config/private.pem");
@@ -462,10 +464,7 @@ async fn global_setup(test_context_builer: TestContextBuilder) -> Result<()> {
 
     // Check if the user would like to skip migrating schemas. This is helpful for boosting
     // performance when running integration tests that do not rely on builtin schemas.
-    let skip_migrating_schemas = match env::var(ENV_VAR_SKIP_MIGRATING_SCHEMAS) {
-        Ok(skip_migrating_schemas_variable) => !skip_migrating_schemas_variable.is_empty(),
-        Err(_) => false,
-    };
+    let builtin_schema_option = assemble_builtin_schema_option();
 
     info!("creating builtins");
     dal::migrate_builtins(
@@ -475,7 +474,7 @@ async fn global_setup(test_context_builer: TestContextBuilder) -> Result<()> {
         services_ctx.veritech().clone(),
         services_ctx.encryption_key(),
         format!("{nats_subject_prefix}.council"),
-        skip_migrating_schemas,
+        builtin_schema_option,
     )
     .await
     .wrap_err("failed to run builtin migrations")?;
@@ -490,4 +489,35 @@ async fn global_setup(test_context_builer: TestContextBuilder) -> Result<()> {
 
     info!("global test setup complete");
     Ok(())
+}
+
+fn assemble_builtin_schema_option() -> BuiltinSchemaOption {
+    match env::var(ENV_VAR_BUILTIN_SCHEMAS) {
+        Ok(found_value) => {
+            let mut builtin_schemas = HashSet::new();
+
+            // If the value does not contain a comma, we will have exactly once item to iterate
+            // over.
+            for builtin_schema in found_value.split(',') {
+                // Trim and ensure the string is lowercase.
+                let cleaned = builtin_schema.trim().to_lowercase();
+
+                // If we receive any keywords indicating that we need to return early, let's do so.
+                if &cleaned == "none" || &cleaned == "false" {
+                    return BuiltinSchemaOption::None;
+                } else if &cleaned == "all" || &cleaned == "true" {
+                    return BuiltinSchemaOption::All;
+                }
+
+                // If we do not find any keywords, we assume that the user provided the name for a
+                // builtin schema.
+                builtin_schemas.insert(cleaned);
+            }
+            BuiltinSchemaOption::Some(builtin_schemas)
+        }
+        Err(_) => {
+            // If the variable is unset, then we migrate everything. This is the default behavior.
+            BuiltinSchemaOption::All
+        }
+    }
 }
