+++ lib/dal/src/builtins/schema/definitions/coreos_butane.json
