+++ lib/dal/src/builtins/schema/docker_hub_credential.rs
@@ -1,0 +1,78 @@
+use crate::schema::variant::leaves::LeafInputLocation;
+use crate::schema::variant::leaves::LeafKind;
+use crate::{builtins::schema::MigrationDriver, schema::variant::leaves::LeafInput};
+use crate::{
+    component::ComponentKind, edit_field::widget::*, socket::SocketArity, BuiltinsError,
+    BuiltinsResult, DalContext, ExternalProvider, Prop, PropKind, SchemaVariant, StandardModel,
+};
+
+impl MigrationDriver {
+    pub async fn migrate_docker_hub_credential(
+        &self,
+        ctx: &DalContext,
+        ui_menu_category: &str,
+        node_color: i64,
+    ) -> BuiltinsResult<()> {
+        let (schema, mut schema_variant, root_prop, _, _, _) = match self
+            .create_schema_and_variant(
+                ctx,
+                "Docker Hub Credential",
+                Some(String::from("Credential")),
+                ui_menu_category,
+                ComponentKind::Credential,
+                Some(node_color),
+                None,
+            )
+            .await?
+        {
+            Some(tuple) => tuple,
+            None => return Ok(()),
+        };
+
+        let mut secret_prop = Prop::new(ctx, "secret", PropKind::Integer, None).await?;
+        secret_prop
+            .set_parent_prop(ctx, root_prop.domain_prop_id)
+            .await?;
+        secret_prop
+            .set_widget_kind(ctx, WidgetKind::SecretSelect)
+            .await?;
+
+        // Qualifications
+        let (qualification_func_id, qualification_func_argument_id) = self
+            .find_func_and_single_argument_by_names(ctx, "si:qualificationDockerHubLogin", "domain")
+            .await?;
+        SchemaVariant::add_leaf(
+            ctx,
+            qualification_func_id,
+            *schema_variant.id(),
+            None,
+            LeafKind::Qualification,
+            vec![LeafInput {
+                location: LeafInputLocation::Domain,
+                func_argument_id: qualification_func_argument_id,
+            }],
+        )
+        .await?;
+
+        let identity_func_item = self
+            .get_func_item("si:identity")
+            .ok_or(BuiltinsError::FuncNotFoundInMigrationCache("si:identity"))?;
+
+        let (_output_provider, _output_socket) = ExternalProvider::new_with_socket(
+            ctx,
+            *schema.id(),
+            *schema_variant.id(),
+            "Docker Hub Credential",
+            None,
+            identity_func_item.func_id,
+            identity_func_item.func_binding_id,
+            identity_func_item.func_binding_return_value_id,
+            SocketArity::Many,
+            false,
+        )
+        .await?;
+
+        schema_variant.finalize(ctx, None).await?;
+        Ok(())
+    }
+}
