+++ lib/dal/src/builtins/schema/systeminit_generic_frame.rs
@@ -3,18 +3,12 @@ use crate::component::ComponentKind;
 use crate::validation::Validation;
 use crate::{BuiltinsResult, ComponentType, DalContext, PropKind, StandardModel};
 
-const FRAME_NODE_COLOR: i64 = 0xFFFFFF;
-
 impl MigrationDriver {
-    pub async fn migrate_systeminit(&self, ctx: &DalContext) -> BuiltinsResult<()> {
-        self.migrate_generic_frame(ctx, "Frames").await?;
-        Ok(())
-    }
-
-    async fn migrate_generic_frame(
+    pub async fn migrate_systeminit_generic_frame(
         &self,
         ctx: &DalContext,
         ui_menu_category: &str,
+        node_color: i64,
     ) -> BuiltinsResult<()> {
         let (schema, mut schema_variant, root_prop, _, _, _) = match self
             .create_schema_and_variant(
@@ -23,7 +17,7 @@ impl MigrationDriver {
                 None,
                 ui_menu_category,
                 ComponentKind::Standard,
-                Some(FRAME_NODE_COLOR),
+                Some(node_color),
                 None,
             )
             .await?
